scraper for indeed fully functional and in lib
